requirements.txt
streamlit
pandas
plotly
numpy
matplotlib
seaborn
SQLAlchemy
python-dateutil
psycopg2-binary
